requirements.txt
langchain-core==0.3.5
langchain-community==0.3.0
langchain-groq==0.2.0
streamlit==1.38.0
python-dotenv==1.0.1
pypdf==5.0.0
bs4==0.0.2
transformers==4.44.2
torch==2.4.1
protobuf==3.20.3
pysqlite3-binary==0.5.3.post1
chromadb==0.3.29
